% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMap.R
\name{binarizeSingleCellData}
\alias{binarizeSingleCellData}
\title{Binarize single cell data}
\usage{
binarizeSingleCellData(object, quantiles = seq(0.15, 0.16, 0.01))
}
\arguments{
\item{object}{A \code{DistMap} object.}

\item{quantiles}{A range of quantiles.}
}
\value{
A \code{DistMap} object with binarized data.
}
\description{
Binarizes the single cell data contained in the \code{DistMap} object. This is done in three steps:
(i) gene thresholds are chosen for a given quantile, (ii) the gene correlations of the binarized
data are compared against those of the \code{insitu.matrix} and a score is computed, (iii) the
binarized data based on the minimal score is computed and added to the \code{DistMap} object.
}
