% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMap.R
\name{computeGeneGradient}
\alias{computeGeneGradient}
\title{Compute gene gradient}
\usage{
computeGeneGradient(object, gene, threshold = 0.75, type = "DV",
  slide.x = 0, slide.y = 0)
}
\arguments{
\item{object}{A \code{DistMap} object.}

\item{gene}{The gene symbol.}

\item{threshold}{A threshold between 0 and 1.}
}
\description{
Computes a gradient of a gene.
}
