% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMap.R
\name{computeMCC}
\alias{computeMCC}
\title{Compute Matthews Correlation Coefficients}
\usage{
computeMCC(object, cell)
}
\arguments{
\item{object}{A \code{DistMap} object.}

\item{cell}{A number denoting the cell.}
}
\value{
The MCC scores between the given cell and all bins.
}
\description{
Computes MCC scores and adds them to the \code{DistMap} object. Requires the \code{binarized.data}
to be computed.
}
