% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMap.R
\name{computeVISH}
\alias{computeVISH}
\title{Compute vISH}
\usage{
computeVISH(object, gene, threshold = 0.75)
}
\arguments{
\item{object}{A \code{DistMap} object.}

\item{gene}{A gene symbol.}

\item{threshold}{A threshold between 0 and 1.}
}
\value{
A vector of gene expression in every bin.
}
\description{
Computes a virtual in situ hybridization for a given gene and a given threshold.
}
