% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMap.R
\name{inSilicoDissect}
\alias{inSilicoDissect}
\title{In Silico Dissect}
\usage{
inSilicoDissect(object, gene.sets)
}
\arguments{
\item{object}{A \code{DistMap} object.}

\item{gene.sets}{A list of gene sets according to which the in silico dissection
will be performed. Each of the gene sets is given as a vector of gene symbols of
minimal length 2.}
}
\value{
A \code{matrix} with scores for every cell against each gene set.
}
\description{
Performs in silico dissection of the embryo according to lisst of genes.
}
