% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMap.R
\name{mapCells}
\alias{mapCells}
\title{Map cells}
\usage{
mapCells(object)
}
\arguments{
\item{object}{A \code{DistMap} object.}
}
\value{
A \code{DistMap} object with computed MCC scores.
}
\description{
Maps single cells onto the reference database.
}
